Extracting styles from logging-selectors
