+++ src/components/logging-selectors/ColorPicker.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 import { Picker } from "@react-native-picker/picker";
-import { StyleSheet, View } from "react-native";
+import { View } from "react-native";
 import { holdOrTapeColors } from "../../constants/holdOrTapeColors";
+import styles from "./styles";
 
 interface ColorProps {
 	color: string;
@@ -24,18 +25,3 @@ export default function ColorPicker({ color, setColor }: ColorProps) {
 		</View>
 	);
 }
-
-const styles = StyleSheet.create({
-	container: {
-		display: "flex",
-		flexDirection: "row",
-		alignItems: "center",
-		justifyContent: "center",
-	},
-	picker: {
-		flex: 2,
-	},
-	pickerItem: {
-		height: 120,
-	},
-});
