+++ src/components/logging-selectors/GradePicker.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 import { Picker } from "@react-native-picker/picker";
-import { StyleSheet, View } from "react-native";
+import { View } from "react-native";
 import { boulderGrades, ropeGrades } from "../../constants/climbingGrades";
+import styles from "./styles";
 
 interface GradeProps {
 	climbingDiscipline: string;
@@ -32,18 +33,3 @@ export default function GradePicker({
 		</View>
 	);
 }
-
-const styles = StyleSheet.create({
-	container: {
-		display: "flex",
-		flexDirection: "row",
-		alignItems: "center",
-		justifyContent: "center",
-	},
-	picker: {
-		flex: 2,
-	},
-	pickerItem: {
-		height: 120,
-	},
-});
