+++ src/components/logging-selectors/PickerContainer.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
-import { StyleSheet, Text, View } from "react-native";
+import styles from "./styles";
+import { Text, View } from "react-native";
 
 type PickerContainerProps = {
 	children: React.ReactNode;
@@ -15,23 +16,3 @@ export default function PickerContainer({ children }: PickerContainerProps) {
 		</View>
 	);
 }
-
-const styles = StyleSheet.create({
-	container: {
-		display: "flex",
-		flexDirection: "row",
-		alignItems: "center",
-		justifyContent: "center",
-	},
-	label: {
-		textAlign: "center",
-		fontSize: 20,
-		flex: 1,
-	},
-	picker: {
-		flex: 2,
-	},
-	pickerItem: {
-		height: 120,
-	},
-});
