+++ src/components/logging-selectors/ProgressPicker.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { Picker } from "@react-native-picker/picker";
-import { StyleSheet, View } from "react-native";
+import { View } from "react-native";
+import styles from "./styles";
 
 interface ProgressProps {
 	progress: string;
@@ -28,18 +29,3 @@ export default function ProgressPicker({
 		</View>
 	);
 }
-
-const styles = StyleSheet.create({
-	container: {
-		display: "flex",
-		flexDirection: "row",
-		alignItems: "center",
-		justifyContent: "center",
-	},
-	picker: {
-		flex: 2,
-	},
-	pickerItem: {
-		height: 120,
-	},
-});
