+++ src/components/logging-selectors/TerrainPicker.tsx
@@ -1,7 +1,8 @@
 import React from "react";
-import { StyleSheet, View } from "react-native";
+import { View } from "react-native";
 import { terrainTypes } from "../../constants/terrainTypes";
 import Checkbox from "../basic-components/Checkbox";
+import styles from "./styles";
 
 interface TerrainProps {
 	terrain: Array<string>
@@ -38,20 +39,3 @@ export default function TerrainPicker({ terrain, setTerrain }: TerrainProps) {
 		</View>
 	);
 }
-
-const styles = StyleSheet.create({
-	container: {
-		alignItems: "center",
-		display: "flex",
-		flexDirection: "row",
-		marginTop: 10,
-	},
-	columnContainer: {
-		display: "flex",
-		flexDirection: "row",
-		gap: 20,
-	},
-	column: {
-		gap: 10,
-	},
-});
