+++ src/components/logging-selectors/index.ts
@@ -11,5 +11,5 @@ export {
 	GradePicker,
 	ProblematicHoldsPicker,
 	ProgressPicker,
-	TerrainPicker
+	TerrainPicker,
 };
